requirements.txt
stegano==0.11.1
Pillow~=9.4.0